backend/requirements.txt
# Django and REST Framework
Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Alith SDK and AI dependencies
alith>=0.12.3

# HTTP client for node executors
httpx>=0.27.0

# Optional AI integrations (uncomment as needed)
# chromadb>=0.4.0
# pymilvus>=2.3.0
# faiss-cpu>=1.7.4
# fastembed>=0.2.0

# Python utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database (for production, consider PostgreSQL)
# psycopg2-binary>=2.9.9

# For async support
asgiref>=3.7.2

